cleanup Acls Manager abstractions
